purchase value to int

+++ Lrr.Shared/IInputData.cs
@@ -9,7 +9,7 @@ namespace Lrr.Shared
     public interface IInputData
     {
         public int WhiteBlackRatio { get; set; }
-        public double PurchaseValue { get; set;}
+        public int PurchaseValue { get; set;}
         public DateOnly RegistrationDate { get; set; }
         public int HighestBidValue { get; set; }
         public DecisionEnum Decision { get; set; }
